requirements.txt
torch
beautifulsoup4 
requests 
lxml
lxml_html_clean
requests_html

yt-dlp 
faster-whisper 
imageio-ffmpeg
openai-whisper

nltk
google-search-results
python-dotenv

openai
flask
trafilatura